Filter a matrix based on expression levels

Filter a matrix based on expression levels

Usage

cleanCtmat(ctmat, geneThr = 0.02, sampleThr = 0.05)

Arguments

ctmat

A numeric matrix where rows represent genes and columns represent samples (e.g., cells).

geneThr

A numeric value between 0 and 1 specifying the minimum proportion of cells in which a gene must be expressed to be retained. The default is 0.02.

sampleThr

A numeric value between 0 and 1 specifying the minimum rank percentile for the number of genes expressed in a sample to retain that sample. The default is 0.05.

Value

The filtered matrix.

Examples

set.seed(123)
ctmat <- matrix(sample(c(0, 1, 2, 3), 30, replace = TRUE),
nrow = 5, ncol = 6,
dimnames = list(
paste0("Gene", 1:5),
paste0("Cell", 1:6)))
cleanCtmat(ctmat)
#>       Cell1 Cell2 Cell3 Cell4 Cell6
#> Gene1     2     1     3     2     2
#> Gene2     2     1     1     3     3
#> Gene3     2     1     1     0     1
#> Gene4     1     2     0     2     2
#> Gene5     2     0     1     2     1