Create a continent column for a dataset with a countryCode column

Create a continent column for a dataset with a countryCode column

Usage

get_continents(data, country = "countryCode", original = "iso3c")

Arguments

data

dataset with observations

country

column name of column with country codes

original

type of data in the source column which contains cuntries to identify continent from

Value

dataset of same type that was passed to it. One column of continents.

Examples

get_continents(obsData)
#>   [1] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#>   [7] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#>  [13] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#>  [19] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#>  [25] "Africa"   "Oceania"  "Oceania"  "Oceania"  "Europe"   "Oceania" 
#>  [31] "Oceania"  "Asia"     "Europe"   "Oceania"  "Oceania"  "Oceania" 
#>  [37] "Americas" "Asia"     "Europe"   "Oceania"  "Oceania"  "Oceania" 
#>  [43] "Oceania"  "Europe"   "Oceania"  "Oceania"  "Americas" "Europe"  
#>  [49] "Oceania"  "Oceania"  "Americas" "Oceania"  "Oceania"  "Europe"  
#>  [55] "Americas" "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#>  [61] "Oceania"  "Europe"   "Asia"     "Europe"   "Asia"     "Americas"
#>  [67] "Oceania"  "Europe"   "Oceania"  "Americas" "Oceania"  "Oceania" 
#>  [73] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Europe"   "Oceania" 
#>  [79] "Americas" "Americas" "Oceania"  "Americas" "Americas" "Oceania" 
#>  [85] "Oceania"  "Europe"   "Oceania"  "Oceania"  "Americas" "Oceania" 
#>  [91] "Oceania"  "Africa"   "Oceania"  "Americas" "Asia"     "Oceania" 
#>  [97] "Europe"   "Asia"     "Oceania"  "Oceania"  "Oceania"  "Asia"    
#> [103] "Oceania"  "Oceania"  "Americas" "Oceania"  "Americas" "Americas"
#> [109] "Americas" "Oceania"  "Oceania"  "Asia"     "Asia"     "Europe"  
#> [115] "Oceania"  "Africa"   "Americas" "Europe"   "Oceania"  "Americas"
#> [121] "Oceania"  "Americas" "Europe"   "Oceania"  "Asia"     "Americas"
#> [127] "Americas" "Oceania"  "Europe"   "Europe"   "Europe"   "Americas"
#> [133] "Oceania"  "Oceania"  "Americas" "Americas" "Asia"     "Europe"  
#> [139] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [145] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [151] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [157] "Europe"   "Americas" "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [163] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Americas" "Asia"    
#> [169] "Europe"   "Europe"   "Asia"     "Oceania"  "Oceania"  "Americas"
#> [175] "Asia"     "Americas" "Oceania"  "Asia"     "Americas" "Oceania" 
#> [181] "Oceania"  "Oceania"  "Europe"   "Oceania"  "Oceania"  "Americas"
#> [187] "Americas" "Americas" "Europe"   "Europe"   "Europe"   "Oceania" 
#> [193] "Oceania"  "Europe"   "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [199] "Oceania"  "Europe"   "Oceania"  "Europe"   "Asia"     "Europe"  
#> [205] "Asia"     "Africa"   "Americas" "Oceania"  "Oceania"  "Europe"  
#> [211] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [217] "Europe"   "Europe"   "Europe"   "Europe"   "Oceania"  "Oceania" 
#> [223] "Oceania"  "Americas" "Americas" "Oceania"  "Oceania"  "Oceania" 
#> [229] "Oceania"  "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [235] "Europe"   "Europe"   "Europe"   "Oceania"  "Oceania"  "Oceania" 
#> [241] "Oceania"  "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [247] "Europe"   "Europe"   "Americas" "Europe"   "Europe"   "Europe"  
#> [253] "Oceania"  "Europe"   "Americas" "Americas" "Europe"   "Europe"  
#> [259] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [265] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [271] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [277] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [283] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [289] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [295] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [301] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [307] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [313] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [319] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [325] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [331] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [337] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [343] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [349] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [355] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [361] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [367] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [373] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [379] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [385] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [391] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [397] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [403] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [409] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [415] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [421] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [427] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [433] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [439] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [445] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [451] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [457] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [463] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [469] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [475] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [481] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [487] "Europe"   "Asia"     "Oceania"  "Americas" "Europe"   "Europe"  
#> [493] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [499] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [505] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [511] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [517] "Oceania"  "Oceania"  "Europe"   "Europe"   "Americas" "Oceania" 
#> [523] "Africa"   "Americas" "Europe"   "Oceania"  "Americas" "Oceania" 
#> [529] "Americas" "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [535] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [541] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [547] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [553] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [559] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [565] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [571] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [577] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [583] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [589] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [595] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [601] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [607] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [613] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [619] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [625] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [631] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [637] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [643] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [649] "Europe"   "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [655] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [661] "Europe"   "Oceania"  "Americas" "Oceania"  "Europe"   "Europe"  
#> [667] "Oceania"  "Oceania"  "Europe"   "Europe"   "Europe"   "Europe"  
#> [673] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [679] "Europe"   "Europe"   "Oceania"  "Oceania"  "Americas" "Oceania" 
#> [685] "Europe"   "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Europe"  
#> [691] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [697] "Europe"   "Europe"   "Asia"     "Oceania"  "Asia"     "Asia"    
#> [703] "Oceania"  "Oceania"  "Asia"     "Oceania"  "Europe"   "Europe"  
#> [709] "Asia"     "Europe"   "Americas" "Oceania"  "Oceania"  "Oceania" 
#> [715] "Oceania"  "Asia"     "Asia"     "Oceania"  "Asia"     "Africa"  
#> [721] "Americas" "Europe"   "Europe"   "Oceania"  "Americas" "Africa"  
#> [727] "Americas" "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [733] "Africa"   "Oceania"  "Asia"     "Europe"   "Oceania"  "Oceania" 
#> [739] "Africa"   "Oceania"  "Americas" "Oceania"  "Oceania"  "Americas"
#> [745] "Europe"   "Americas" "Oceania"  "Oceania"  "Americas" "Asia"    
#> [751] "Oceania"  "Europe"   "Oceania"  "Americas" "Europe"   "Europe"  
#> [757] "Europe"   "Europe"   "Europe"   "Europe"   "Africa"   "Asia"    
#> [763] "Europe"   "Asia"     "Americas" "Europe"   "Europe"   "Americas"
#> [769] "Oceania"  "Asia"     "Oceania"  "Europe"   "Oceania"  "Americas"
#> [775] "Asia"     "Europe"   "Americas" "Americas" "Asia"     "Oceania" 
#> [781] "Oceania"  "Europe"   "Americas" "Asia"     "Americas" "Asia"    
#> [787] "Europe"   "Americas" "Asia"     "Americas" "Americas" "Asia"    
#> [793] "Americas" "Oceania"  "Americas" "Europe"   "Europe"   "Europe"  
#> [799] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [805] "Europe"   "Oceania"  "Europe"   "Americas" "Europe"   "Americas"
#> [811] "Americas" "Oceania"  "Americas" "Americas" "Americas" "Europe"  
#> [817] "Europe"   "Europe"   "Europe"   "Americas" "Americas" "Europe"  
#> [823] "Europe"   "Americas" "Americas" "Americas" "Americas" "Europe"  
#> [829] "Europe"   "Africa"   "Americas" "Americas" "Americas" "Europe"  
#> [835] "Europe"   "Europe"   "Americas" "Americas" "Europe"   "Americas"
#> [841] "Oceania"  "Americas" "Europe"   "Europe"   "Europe"   "Europe"  
#> [847] "Europe"   "Asia"     "Americas" "Europe"   "Europe"   "Americas"
#> [853] "Americas" "Americas" "Oceania"  "Europe"   "Oceania"  "Europe"  
#> [859] "Americas" "Oceania"  "Europe"   "Asia"     "Asia"     "Europe"  
#> [865] "Asia"     "Asia"     "Asia"     "Oceania"  "Americas" "Europe"  
#> [871] "Oceania"  "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [877] "Europe"   "Europe"   "Europe"   "Europe"   "Americas" "Americas"
#> [883] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [889] "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania"  "Oceania" 
#> [895] "Americas" "Europe"   "Oceania"  "Americas" "Oceania"  "Oceania" 
#> [901] "Americas" "Oceania"  "Americas" "Europe"   "Europe"   "Europe"  
#> [907] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [913] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [919] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [925] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [931] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [937] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [943] "Europe"   "Europe"   "Europe"   "Europe"   "Asia"     "Americas"
#> [949] "Asia"     "Asia"     "Asia"     "Asia"     "Asia"     "Europe"  
#> [955] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [961] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [967] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [973] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [979] "Europe"   "Europe"   "Europe"   "Europe"   "Europe"   "Europe"  
#> [985] "Europe"