Generate cell type specific matrices with barcodes

Generate cell type specific matrices with barcodes

Usage

generate_cell_types_w_barcodes(cell_types, sample_list, cell_IDs)

Arguments

cell_types

Vector of cell types to process (e.g., c("T_cells", "B_cells")).

sample_list

List of matrices where rows are genes and columns are cells. Each matrix should have column names indicating cell types. Genes should be the same (same number and order) for all matrices.

cell_IDs

A dataframe that should have two columns: "cell_type" and "cell_ID".

Value

A list of sparse matrices (one for each cell type in cell_types), with cell IDs as column names and shared row names.

Examples

cell_IDs <- data.frame(
cell_type = c("T_cells", "B_cells", "T_cells", "B_cells",
"T_cells", "T_cells", "B_cells", "B_cells"),
cell_ID = c("Cell1", "Cell2", "Cell3", "Cell4",
"Cell5", "Cell6", "Cell7", "Cell8"))

sample1 <- matrix(
c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12),
nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"),
c("Cell1", "Cell2", "Cell3", "Cell4")))

sample2 <- matrix(
c(13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24),
nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"),
c("Cell5", "Cell6", "Cell7", "Cell8")))
sample_list <- list(sample1, sample2)
cell_types <- c("T_cells", "B_cells")
generate_cell_types_w_barcodes(cell_types, sample_list, cell_IDs)
#> $T_cells
#> 3 x 4 sparse Matrix of class "dgCMatrix"
#>       Cell1 Cell3 Cell5 Cell6
#> Gene1     1     7    13    16
#> Gene2     2     8    14    17
#> Gene3     3     9    15    18
#> 
#> $B_cells
#> 3 x 4 sparse Matrix of class "dgCMatrix"
#>       Cell2 Cell4 Cell7 Cell8
#> Gene1     4    10    19    22
#> Gene2     5    11    20    23
#> Gene3     6    12    21    24
#>