Load dataset with new column names

Given an URL of an online dataset, the datasets' delimiters, and a vector of new column names, loads the dataset as a tibble with the new column names.

Usage

load_w_new_col(dataset_path, col_names, delimiter)

Arguments

dataset_path

An URL or path that leads to a dataset.

col_names

Vector with the new column names as strings (quoted).

delimiter

The dataset's delimiter (could be ".", ";", ",", etc), quoted.

Value

Loaded dataset with the new column names, as a tibble.

Examples

load_w_new_col("https://raw.githubusercontent.com/plotly/datasets/master/mtcars.csv",
c("model", "milespergallon", "cylinder_number",
"displacement", "horsepower", "rear_axle_ratio",
"weight", "quartermiletime", "engine", "transmission",
"forwardgears", "carb"), ",")
#> Rows: 32 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (1): manufacturer
#> dbl (11): mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 32 × 12
#>    model  milespergallon cylinder_number displacement horsepower rear_axle_ratio
#>    <chr>           <dbl>           <dbl>        <dbl>      <dbl>           <dbl>
#>  1 Mazda…           21                 6         160         110            3.9 
#>  2 Mazda…           21                 6         160         110            3.9 
#>  3 Datsu…           22.8               4         108          93            3.85
#>  4 Horne…           21.4               6         258         110            3.08
#>  5 Horne…           18.7               8         360         175            3.15
#>  6 Valia…           18.1               6         225         105            2.76
#>  7 Duste…           14.3               8         360         245            3.21
#>  8 Merc …           24.4               4         147.         62            3.69
#>  9 Merc …           22.8               4         141.         95            3.92
#> 10 Merc …           19.2               6         168.        123            3.92
#> # ℹ 22 more rows
#> # ℹ 6 more variables: weight <dbl>, quartermiletime <dbl>, engine <dbl>,
#> #   transmission <dbl>, forwardgears <dbl>, carb <dbl>